+++ signals_windows.go
@@ -1,0 +1,7 @@
+// +build windows
+
+package main
+
+func setupSignals() {
+	// NOOP: Windows doesn't have signals equivalent to the Unix world.
+}
