+++ webhook.go
@@ -1,5 +1,3 @@
-//+build !windows
-
 package main
 
 import (
@@ -12,9 +10,7 @@ import (
 	"net/url"
 	"os"
 	"os/exec"
-	"os/signal"
 	"strings"
-	"syscall"
 
 	"github.com/adnanh/webhook/hook"
 
@@ -61,12 +57,7 @@ func init() {
 	log.Println("version " + version + " starting")
 
 	// set os signal watcher
-	log.Printf("setting up os signal watcher\n")
-
-	signals = make(chan os.Signal, 1)
-	signal.Notify(signals, syscall.SIGUSR1)
-
-	go watchForSignals()
+	setupSignals()
 
 	// load and parse hooks
 	log.Printf("attempting to load hooks from %s\n", *hooksFilePath)
@@ -281,21 +272,6 @@ func watchForFileChange() {
 			}
 		case err := <-(*watcher).Errors:
 			log.Println("watcher error:", err)
-		}
-	}
-}
-
-func watchForSignals() {
-	log.Println("os signal watcher ready")
-
-	for {
-		sig := <-signals
-		if sig == syscall.SIGUSR1 {
-			log.Println("caught USR1 signal")
-
-			reloadHooks()
-		} else {
-			log.Printf("caught unhandled signal %+v\n", sig)
 		}
 	}
 }
